Python/New/MCP-Client/requirements.txt
dotenv==0.9.9
azure-identity==1.25.1
azure-ai-projects==2.0.0b3
openai==2.14.0